refactor: decompose large BatchJobTable component into smaller subcomponents
Extract subcomponents from the monolithic `BatchJobTable.tsx` file to improve
readability and adherence to SOLID principles. The table headers, rows,
pagination, and state components (loading, empty, error) are now
standalone components within a new `table/` directory.

All related tests have been verified to pass.
